modify,test - 주석처리, test code 작성 및 정리(#22)
view 파일 주석 정리, 로그인, 로그아웃 테스트코드 추가 및 주석 처리
